backend/ai_services/requirements.txt
transformers>=4.30.0
torch>=2.0.0
spacy>=3.6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
textstat>=0.7.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.1.0

# File processing dependencies
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
mammoth>=1.6.0
chardet>=5.0.0

# AI model dependencies
bitsandbytes>=0.41.0
accelerate>=0.20.0 